Use gorilla mux middleware application method
Gorilla mux will compose and apply middleware for us. This is clearer, since the order of execution is well defined (order of application). It also will help avoid lost of composed constructor calls if we need multiple middleware. It's an extra line of code though, and those subrouters are required to make it work.